Add files via upload
A model for diagnosis of  diabetes was developed using Artificial Neural Network (ANN) method . In this study, a hybrid ANN-Genetic Algorithm model was developed for classification of diabetic patients. Therefore, the number of optimal neurons as well as hidden layers was determined to design the architecture of the ANN model. In this study, a model for diagnosis of  diabetes was developed using Artificial Neural Network (ANN) method.